Changed a background
Changed the background on the #topsection id to something more appropriate for what we're doing (Used the red & white basically). May need some tweaking as the logo fades into it a bit, but other than that i personally think it looks better. Of course, we need to change the background of the main bit too but I couldn't find the line for it. Let me know what you think anyways!